[Refactor] component 패키지명 widget으로 변경
